+++ contracts/mocks/ERC20Mock.sol
@@ -4,7 +4,7 @@ import "../token/ERC20/ERC20.sol";
 
 // mock class using ERC20
 contract ERC20Mock is ERC20 {
-    constructor (address initialAccount, uint256 initialBalance) public {
+    constructor (address initialAccount, uint256 initialBalance) public payable {
         _mint(initialAccount, initialBalance);
     }